requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
aiohttp>=3.8.0
python-multipart>=0.0.6
requests>=2.28.0
Pillow>=9.0.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
git+https://github.com/openai/CLIP.git
litserve>=0.2.13
pandas>=1.5.0